// src/app/Interfaces/ispace-item.ts
export interface IspaceItem {

    title: string;
    description: string;
    rentPerMonth: number;
    maxRequierd: number;
    currenltyPresent: number;
    university: string;
    address1: string;
    address2: string;
    city: string;
    state: string;
    zipCode: string;
    spaceId :string;
    ownerId:string;

}